requirements.txt
streamlit==1.43.2
pandas==2.2.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.115.0
ollama==0.4.7
pydantic==2.9.2
agno==0.5.0
psutil==5.9.6